requirements.txt
python-telegram-bot==20.3
transformers==4.30.2
torch==2.0.1
fastapi==0.104.1
uvicorn==0.24.0
redis==4.5.5
python-multipart==0.0.6
librosa==0.10.0
gunicorn==20.1.0
numpy<2
